Move watcher into bin file
